Создать toolbar в активити "Список покупок"
Тулбар с надписью "Список покупок" и кнопкой возврата.
