+++ app/src/main/res/layout/activity_main3.xml
@@ -21,7 +21,8 @@
 
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/content_main3" />
+    <include layout="@layout/content_main3"
+        android:id="@+id/toolbar" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
