+++ app/src/main/res/layout/activity_spisok_produktov.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="ru.SpisokProduktov.SpisokProduktov">
+    tools:context="i.layout.SpisokProduktov">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
