+++ app/src/main/res/layout/content_spisok_produktov.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="ru.SpisokProduktov.SpisokProduktov"
+    tools:context="i.layout.SpisokProduktov"
     tools:showIn="@layout/activity_spisok_produktov">
 
 </RelativeLayout>
